Screenshots: Hoist multi-language uiStrings into integration-test-utils
The VxMark and VxScan screenshot suites carried the same hand-written
multiLangBallotStrings table (patched into the election's ballotStrings) that
the prior commit removed from VxMarkScan. Apply the same fix and hoist the
shared piece into @votingworks/integration-test-utils so all three suites use
one source.

- Add MULTI_LANGUAGE_UI_STRINGS to integration-test-utils, generating the
  ballotLanguage names via format.languageDisplayName (Intl) and registering
  them through the election package's uiStrings, so the voter language selector
  renders each language in its own script (English, español, 简体中文, 繁體中文).
- mark-scan/mark/scan now import the shared constant instead of defining their
  own; mark/scan drop the ballotStrings patching and (for scan) the inline
  table. App-specific election builders stay local.
