+++ apps/mark-scan/integration-testing/package.json
@@ -22,7 +22,6 @@
     "@votingworks/integration-test-utils": "workspace:*",
     "@votingworks/types": "workspace:*",
     "@votingworks/usb-drive": "workspace:*",
-    "@votingworks/utils": "workspace:*",
     "concurrently": "7.6.0",
     "dotenv": "16.3.1"
   },
