+++ apps/mark/integration-testing/e2e/screenshots.spec.ts
@@ -7,6 +7,7 @@ import {
   setupTemporaryRootDir,
 } from '@votingworks/fixtures';
 import {
+  MULTI_LANGUAGE_UI_STRINGS,
   buildIntegrationTestHelper,
   captureReadinessReport,
   createScreenshotCounter,
@@ -28,9 +29,8 @@ import {
   logInAsSystemAdministrator,
 } from './support/auth';
 import {
-  getMultiLanguageFamousNamesElectionDefinition,
+  getFamousNamesElectionDefinition,
   getMultiLanguageGeneralElectionDefinition,
-  MULTI_LANGUAGE_UI_STRINGS,
 } from './support/election';
 import { configureMachine, openPolls, voteFullBallot } from './support/flows';
 import { capturePrintedBallot } from './support/reports';
@@ -62,8 +62,8 @@ async function buildElectionPackage(electionDefinition: ElectionDefinition) {
       // Hide the voter help button to keep voter screens uncluttered.
       disableVoterHelpButtons: true,
     },
-    // Registers the language codes; the display names come from the election's
-    // ballotStrings (see getMultiLanguageFamousNamesElectionDefinition).
+    // Registers the supported languages and their native display names so the
+    // voter language selector renders (see MULTI_LANGUAGE_UI_STRINGS).
     uiStrings: MULTI_LANGUAGE_UI_STRINGS,
   });
 }
@@ -145,7 +145,7 @@ async function voteAndCaptureContests(
 }
 
 test('basic election flow', async ({ page }) => {
-  const electionDefinition = getMultiLanguageFamousNamesElectionDefinition();
+  const electionDefinition = getFamousNamesElectionDefinition();
   const { election } = electionDefinition;
   const usbHandler = getMockFileUsbDriveHandler();
   const helper = buildIntegrationTestHelper(page, screenshotCounter);
@@ -313,7 +313,7 @@ test('basic election flow', async ({ page }) => {
 });
 
 test('additional options', async ({ page }) => {
-  const electionDefinition = getMultiLanguageFamousNamesElectionDefinition();
+  const electionDefinition = getFamousNamesElectionDefinition();
   const { election } = electionDefinition;
   const helper = buildIntegrationTestHelper(page, screenshotCounter);
   const {
