+++ libs/integration-test-utils/src/index.ts
@@ -5,3 +5,4 @@ export * from './screenshots';
 export * from './pdf';
 export * from './playwright_config';
 export * from './reports';
+export * from './ui_strings';
